requirements.txt
flask
gunicorn
-e git+https://github.com/SimonSapin/Frozen-Flask.git#egg=frozen-flask
Flask-Assets
hamlish-jinja
python-frontmatter
mailchimp
python-dotenv
logging
pyopenssl
ndg-httpsclient
pyasn1
urllib3